+++ Changes.md
@@ -32,6 +32,7 @@ API
 - OpenColorIOTransformUI :
   - Added `noneLabel` argument to `colorSpacePresetNames()`.
   - Added support for `openColorIO:categories` and `openColorIO:includeRoles` metadata to `colorSpacePresetNames()`. These may be registered on a per-plug basis to control the colorspaces shown for that plug.
+- ValuePlugs : Added Python bindings for `ValueType` type alias.
 
 Breaking Changes
 ----------------